A python tool to thin/filter on compilations databases
Initial version of a python tool which is capable of producing
a thinned-out compilations database, given a list of (changed) source files
and or a list of (changed) header files.

The primary goal of this tool is to accelerate analysis tasks (static code analysis)
based on compilations database by only focusing on stuff that might have changed
(for instance in a pull request).